Modify SessionGuard to support decimal MaxContinuousUptime and service startup-based session tracking
- Change MaxContinuousUptime from int to double to support fractional hours (e.g., 0.02 for 72 seconds)
- Update SessionInfo.Initialize() to use service startup time as reference instead of Windows session logon time
- Implement cumulative uptime tracking across multiple logins within the same day
- Ensure service continues running after logout and re-initializes SessionManager for new session detection
- Add .gitignore entry for publish/ directory

+++ SessionGuard/SessionConfig.cs
@@ -34,11 +34,11 @@ public class SessionConfig
     public List<LogoutTimeWindow> LogoutTimeWindows { get; set; } = new();
 
     /// <summary>
-    /// 最大連続稼働時間（時間単位）
+    /// 最大連続稼働時間（時間単位、小数対応）
     /// この時間を超えて稼働している場合、ログアウトする
-    /// 例: 1 (1時間)
+    /// 例: 1 (1時間), 0.5 (30分), 1.5 (1時間30分)
     /// </summary>
-    public int MaxContinuousUptime { get; set; } = 8;
+    public double MaxContinuousUptime { get; set; } = 8;
 
     /// <summary>
     /// ログアウト実行を有効にするかどうか
